0x02:task(102) - corrected issue

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,12 +8,12 @@
 int main(void)
 {
 	int count = 0; 
-	unsigned long long a = 0, b = 1, c = 0;
+	unsigned long a = 0, b = 1, c = 0;
 
 	while (count < 50)
 	{
 		c = a + b;
-		printf("%llu", c);
+		printf("%lu", c);
 		a = b;
 		b = c;
 		if (count < 49)
